% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FUN_kernels}
\alias{FUN_kernels}
\title{performs kernel smoothing using a bandwidth. Besides using a kernel there is also the option to combine kernels}
\usage{
FUN_kernels(kernel, W, h)
}
\description{
performs kernel smoothing using a bandwidth. Besides using a kernel there is also the option to combine kernels
}
\keyword{internal}

