% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_categorical_preds}
\alias{func_categorical_preds}
\title{OPTION to convert categorical features TO either numeric [ if levels more than 32] OR to dummy variables [ if levels less than 32 ]}
\usage{
func_categorical_preds(prepr_categ)
}
\description{
OPTION to convert categorical features TO either numeric [ if levels more than 32] OR to dummy variables [ if levels less than 32 ]
}
\keyword{internal}
