% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_tbl}
\alias{func_tbl}
\title{this function returns a table of probabilities for each label}
\usage{
func_tbl(DF, W, labels)
}
\description{
this function returns a table of probabilities for each label
}
\keyword{internal}

