% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_tbl_dist}
\alias{func_tbl_dist}
\title{this function returns the probabilities in case of classification}
\usage{
func_tbl_dist(DF, Levels)
}
\description{
this function returns the probabilities in case of classification
}
\keyword{internal}

