% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FUNCTION_weights}
\alias{FUNCTION_weights}
\title{this function is used as a kernel-function-identifier [ takes the distances and a weights-kernel (in form of a function) and returns weights ]}
\usage{
FUNCTION_weights(W_dist_matrix, weights_function, eps = 1e-06)
}
\description{
this function is used as a kernel-function-identifier [ takes the distances and a weights-kernel (in form of a function) and returns weights ]
}
\keyword{internal}

