% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regr_folds}
\alias{regr_folds}
\title{create folds (in regression)                                           [ detailed information about class_folds in the FeatureSelection package ]}
\usage{
regr_folds(folds, RESP)
}
\description{
this function creates both stratified and non-stratified folds in regression
}
\keyword{internal}

