% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch.ops}
\alias{switch.ops}
\title{Arithmetic operations on lists}
\usage{
switch.ops(LST, MODE = "ADD")
}
\description{
Arithmetic operations on lists
}
\references{
https://www.cs.toronto.edu/~duvenaud/cookbook/

https://raw.githubusercontent.com/duvenaud/phd-thesis/master/kernels.pdf
}
\keyword{internal}
